Move `can_be_cropped_to?` into `picture_thumbnails`
We only need it for `allow_image_cropping?`
