+++ app/models/alchemy/picture/transformations.rb
@@ -7,10 +7,6 @@ module Alchemy
   module Picture::Transformations
     extend ActiveSupport::Concern
 
-    included do
-      include Alchemy::Picture::Calculations
-    end
-
     # Returns the rendered cropped image. Tries to use the crop_from and crop_size
     # parameters. When they can't be parsed, it just crops from the center.
     #
@@ -109,6 +105,30 @@ module Alchemy
       {
         width: [dimensions[:width].to_i, image_file_width.to_i].min,
         height: [dimensions[:height].to_i, image_file_height.to_i].min
+      }
+    end
+
+    # Returns true if both dimensions of the base image are bigger than the dimensions hash.
+    #
+    def is_bigger_than?(dimensions)
+      image_file_width > dimensions[:width] && image_file_height > dimensions[:height]
+    end
+
+    # Returns true is one dimension of the base image is smaller than the dimensions hash.
+    #
+    def is_smaller_than?(dimensions)
+      !is_bigger_than?(dimensions)
+    end
+
+    # Given a string with an x, this function returns a Hash with point
+    # :width and :height.
+    #
+    def sizes_from_string(string)
+      width, height = string.to_s.split("x", 2).map(&:to_i)
+
+      {
+        width: width,
+        height: height
       }
     end
   end
