+++ lib/alchemy/test_support/having_picture_thumbnails_examples.rb
@@ -608,10 +608,15 @@ RSpec.shared_examples_for "having picture thumbnails" do
     let(:picture) { Alchemy::Picture.new }
     let(:image_file_width) { 400 }
     let(:image_file_height) { 300 }
+    let(:crop_size) { "400x300" }
+    let(:upsample) { false }
 
     before do
       allow(picture).to receive(:image_file_width) { image_file_width }
       allow(picture).to receive(:image_file_height) { image_file_height }
+      allow(record).to receive(:settings) do
+        {crop: true, size: crop_size, upsample: upsample}
+      end
     end
 
     subject { record.allow_image_cropping? }
@@ -623,23 +628,37 @@ RSpec.shared_examples_for "having picture thumbnails" do
         allow(record).to receive(:picture) { picture }
       end
 
-      it { is_expected.to be_falsy }
+      context "and image smaller or equal to crop size" do
+        context "if picture.image_file is nil" do
+          before do
+            expect(picture.image_file).to receive(:attached?) { false }
+          end
 
-      context "and with image larger than crop size" do
-        before do
-          allow(picture).to receive(:can_be_cropped_to?) { true }
+          it { is_expected.to be_falsy }
         end
 
+        context "if picture.image_file is present" do
+          before do
+            expect(picture.image_file).to receive(:attached?) { true }
+          end
+
+          it { is_expected.to be_falsy }
+
+          context "but with upsample set to true" do
+            let(:upsample) { true }
+
+            it { is_expected.to be(true) }
+          end
+        end
+      end
+
+      context "and with image larger than crop size" do
         let(:image_file_width) { 1201 }
         let(:image_file_height) { 481 }
 
         it { is_expected.to be_falsy }
 
         context "with crop set to true" do
-          before do
-            allow(record).to receive(:settings) { {crop: true} }
-          end
-
           context "if picture.image_file is nil" do
             before do
               expect(picture.image_file).to receive(:attached?) { false }
