+++ spec/rails_helper.rb
@@ -33,7 +33,6 @@ require "alchemy/test_support/shared_link_tab_examples"
 require "alchemy/test_support/shared_uploader_examples"
 require "alchemy/test_support/current_language_shared_examples"
 
-require_relative "support/calculation_examples"
 require_relative "support/hint_examples"
 require_relative "support/custom_news_elements_finder"
